requirements.txt
Flask==0.12
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
Jinja2==2.9.4
Mako==1.0.7
MarkupSafe==0.23
PyJWT==1.4.2
SQLAlchemy==1.2.7
WTForms==2.1
Werkzeug==0.11.15
alembic==0.9.9
argparse==1.2.1
click==6.7
gunicorn==19.6.0
itsdangerous==0.24
psycopg2==2.7.4
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
wsgiref==0.1.2
